fix: infer "config.release" automatically if not provided (#205)
Use Sentry CLI to automatically deduce "release" configuration if not
specified manually.

It's necessary to store the release in a generated file at build-time,
and import that file at run-time, so that the server-side Sentry can use
it on initialization (which happens within the module itself rather than
within a server-side plugin).

Use hooks to determine when to store and read server-side settings.
Use "build:before" hook to determine release (if not set manually).
Use "listen" hook to read server-side config and initialize it (if enabled).
