+++ .eslintignore
@@ -1,4 +1,5 @@
 !.eslintrc.js
 !.release-it.js
 lib/plugin.*
+lib/templates/*.*
 test/fixture/*/app
